requirements.txt
tk==0.1.0
mysql-connector-python==8.0.32